ir+metainterp: move Box/BoxKind/BoxRef/Forwarded to majit-ir::box_ref
Box, BoxKind, Forwarded, BoxRef, PtrInfoBorrow, PtrInfoBorrowMut,
IntBoundBorrow, IntBoundBorrowMut now live in majit-ir. Dependencies
(OpInfo / PtrInfo / IntBound / Value / Type / OpRef) are all majit-ir-local.

majit-metainterp::box re-exports BoxRef / Forwarded / PtrInfoBorrowMut so
existing callers via crate::r#box::* continue to compile. BoxPool remains
in metainterp because it embeds the pyre-only flat OpRef→Box mapping
that has no PyPy analogue.

The 26 cfg(test) cases that exercised BoxRef/Forwarded chain walking and
ptr_info/int_bound borrow helpers move to majit-ir::box_ref's test module,
where they import IntBound/OpInfo/PtrInfo directly from majit-ir. A new
BoxRef::as_ptr() helper exposes raw identity for the eq-by-pointer test
without exposing the inner Rc field.
